Don't emit ANSI colour escape sequences when output is not a tty
When piping the output of utest.h into a file or pipe the default
behaviour should be to omit colour formatting as these output devices
are non-interactive and colour output confounds parsing.

In a future patch I'd like to add support to make this optional a'la

git diff --color=<auto|always|never> or gcc's
`-fdiagnostics-color=<auto|always|never>`

However, this should be sufficient for now.